tui: extract chatwidget service tier helpers
